# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "药房进销存与处方调配后端服务",
        "title": "药房管理系统 API",
        "contact": {},
        "version": "1.3.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/class-interactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "class-interactions"
                ],
                "summary": "分类交互规则列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "class-interactions"
                ],
                "summary": "新建分类交互规则",
                "parameters": [
                    {
                        "description": "分类交互规则",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.ClassInteractionRule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/class-interactions/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "class-interactions"
                ],
                "summary": "更新分类交互规则",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "分类交互规则",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.ClassInteractionRule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "class-interactions"
                ],
                "summary": "删除分类交互规则",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
        "/drug-ingredients/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "drug-ingredients"
                ],
                "summary": "删除药品成分",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "成分映射ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
        "/drugs/{id}/ingredients": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "drug-ingredients"
                ],
                "summary": "药品成分列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "药品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "drug-ingredients"
                ],
                "summary": "添加药品成分",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "药品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "成分信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.DrugIngredient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
        "/ingredient-interactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "ingredient-interactions"
                ],
                "summary": "成分交互规则列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ingredient-interactions"
                ],
                "summary": "新建成分交互规则",
                "parameters": [
                    {
                        "description": "成分交互规则",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.IngredientInteraction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/ingredient-interactions/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ingredient-interactions"
                ],
                "summary": "更新成分交互规则",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "成分交互规则",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.IngredientInteraction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "ingredient-interactions"
                ],
                "summary": "删除成分交互规则",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
        "/tag-interactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "tag-interactions"
                ],
                "summary": "标签交互规则列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tag-interactions"
                ],
                "summary": "新建标签交互规则",
                "parameters": [
                    {
                        "description": "标签交互规则",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.TagInteraction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/tag-interactions/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tag-interactions"
                ],
                "summary": "更新标签交互规则",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "标签交互规则",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.TagInteraction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "tag-interactions"
                ],
                "summary": "删除标签交互规则",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_handler.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "yaofang_internal_model.ClassInteractionRule": {
            "type": "object",
            "properties": {
                "class_a": {
                    "type": "string"
                },
                "class_b": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "evidence_level": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "level": {
                    "type": "integer"
                },
                "mechanism": {
                    "type": "string"
                },
                "source_reference": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "yaofang_internal_model.DrugIngredient": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "drug_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "ingredient_name": {
                    "type": "string"
                },
                "strength": {
                    "type": "string"
                }
            }
        },
...
        "yaofang_internal_model.IngredientInteraction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "evidence_level": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ingredient_a": {
                    "type": "string"
                },
                "ingredient_b": {
                    "type": "string"
                },
                "level": {
                    "description": "1禁忌 2慎用 3注意",
                    "type": "integer"
                },
                "mechanism": {
                    "type": "string"
                },
                "source_reference": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "yaofang_internal_model.TagInteraction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "evidence_level": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "level": {
                    "type": "integer"
                },
                "mechanism": {
                    "type": "string"
                },
                "source_reference": {
                    "type": "string"
                },
                "tag_a": {
                    "type": "string"
                },
                "tag_b": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}